# internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Garmr API",
    "description": "CUE-based policy evaluation engine for enforcing governance, security, and compliance policies.",
    "version": "1.0.0",
    "contact": {
      "name": "Garmr",
      "url": "https://github.com/infrashift/garmr"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/v1/evaluate": {
      "post": {
        "tags": ["evaluation"],
        "summary": "Evaluate input against policies",
        "description": "Evaluates the provided input against loaded policies and returns the decision with any violations",
        "operationId": "evaluate",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Request ID for audit correlation (generated if not provided)",
            "required": false,
            "schema": {
              "type": "string",
              "example": "pipeline-12345"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              },
              "examples": {
                "kubernetes-pod": {
                  "summary": "Kubernetes Pod evaluation",
                  "value": {
                    "input": {
                      "kind": "Pod",
                      "metadata": {
                        "name": "web-app",
                        "namespace": "production"
                      },
                      "spec": {
                        "containers": [
                          {
                            "name": "app",
                            "image": "gcr.io/my-project/app:v1.0.0"
                          }
                        ]
                      }
                    },
                    "namespace": "security"
                  }
                },
                "release-artifact": {
                  "summary": "Release artifact evaluation",
                  "value": {
                    "input": {
                      "kind": "release",
                      "version": "2.1.0",
                      "tests": {
                        "passed": true,
                        "coverage": 85
                      }
                    },
                    "namespace": "release"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evaluation completed successfully",
            "headers": {
              "X-Request-Id": {
                "description": "Request ID (echoed or generated)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                },
                "examples": {
                  "allow": {
                    "summary": "Policy passed",
                    "value": {
                      "decision": "allow",
                      "request_id": "abc-123",
                      "results": [],
                      "metrics": {
                        "evaluation_time_ns": 1234567,
                        "policies_evaluated": 2,
                        "rules_evaluated": 8
                      }
                    }
                  },
                  "deny": {
                    "summary": "Policy failed",
                    "value": {
                      "decision": "deny",
                      "request_id": "abc-123",
                      "results": [
                        {
                          "policy_name": "container-security",
                          "policy_namespace": "security",
                          "rule_id": "SEC-001",
                          "description": "Containers must not run as root",
                          "severity": "high",
                          "passed": false,
                          "message": "Container is running as root"
                        }
                      ],
                      "metrics": {
                        "evaluation_time_ns": 5234567,
                        "policies_evaluated": 2,
                        "rules_evaluated": 8
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluateRequest": {
        "type": "object",
        "properties": {
          "input": {
            "type": "object",
            "description": "The resource to evaluate",
            "additionalProperties": true
          },
          "namespace": {
            "type": "string",
            "description": "Policy namespace filter",
            "example": "security"
          },
          "policies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Specific policies to evaluate (namespace/name format)",
            "example": ["security/container-security"]
          },
          "trace": {
            "type": "boolean",
            "description": "Include evaluation trace",
            "default": false
          },
          "include_passed": {
            "type": "boolean",
            "description": "Include passed rules in results",
            "default": false
          },
          "strict": {
            "type": "boolean",
            "description": "Treat warnings as failures",
            "default": false
          }
        },
        "required": ["input"]
      },
      "EvaluateResponse": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": ["allow", "deny", "warn"],
            "description": "Overall policy decision"
          },
          "request_id": {
            "type": "string",
            "description": "Request ID for audit correlation"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleResult"
            },
            "description": "Individual rule evaluation results"
          },
          "metrics": {
            "$ref": "#/components/schemas/EvaluationMetrics"
          }
        },
        "required": ["decision", "request_id", "results", "metrics"]
      },
      "RuleResult": {
        "type": "object",
        "properties": {
          "policy_name": {
            "type": "string",
            "description": "Name of the policy"
          },
          "policy_namespace": {
            "type": "string",
            "description": "Namespace of the policy"
          },
          "rule_id": {
            "type": "string",
            "description": "Unique rule identifier"
          },
          "description": {
            "type": "string",
            "description": "Rule description"
          },
          "severity": {
            "type": "string",
            "enum": ["critical", "high", "medium", "low", "info"],
            "description": "Rule severity"
          },
          "passed": {
            "type": "boolean",
            "description": "Whether the rule passed"
          },
          "message": {
            "type": "string",
            "description": "Failure message (if rule failed)"
          }
        },
        "required": ["policy_name", "policy_namespace", "rule_id", "severity", "passed"]
      },
      "EvaluationMetrics": {
        "type": "object",
        "properties": {
          "evaluation_time_ns": {
            "type": "integer",
            "format": "int64",
            "description": "Evaluation time in nanoseconds"
          },
          "policies_evaluated": {
            "type": "integer",
            "description": "Number of policies evaluated"
          },
          "rules_evaluated": {
            "type": "integer",
            "description": "Number of rules evaluated"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "description": "Additional error details"
          }
        },
        "required": ["error"]
      }
    }
  }
}